refactor(domain): 统一使用UUID作为实体ID类型
- 将LoginUserUseCase中用户ID类型由Int改为UUID，修正addUser返回值赋值
- SaveLoginSessionUseCase中userId参数类型改为UUID，调整相关调用
- AddCourseUseCase、AddOrdinaryScheduleUseCase、GetOrdinaryScheduleByIdUseCase、GetOrdinarySchedulesUseCase中相关ID类型改为UUID
- GetUserProfileUseCase和SaveAvatarUseCase中获取用户ID时去除toInt转换，直接使用UUID
- GetDefaultTableTimeConfigUseCase中tableId参数类型改为UUID，调整无效ID判断逻辑
- NavGraph添加@RequiresApi注解，指定最低API版本为VANILLA_ICE_CREAM

此改动统一了项目中实体ID的类型，提升类型安全性，避免了Int与UUID混用导致的潜在错误。

+++ app/src/main/java/com/example/todoschedule/domain/use_case/auth/LoginUserUseCase.kt
@@ -52,7 +52,7 @@ class LoginUserUseCase @Inject constructor(
                     } else {
                         // 本地没有该用户，创建新用户
                         val userId = userRepository.addUser(remoteUser)
-                        return Result.success(remoteUser.copy(id = userId.toInt()))
+                        return Result.success(remoteUser.copy(id = userId))
                     }
                 }
                 // 如果远程登录失败，但不是网络错误等严重问题，可以尝试本地登录
